fix test

+++ mmaction/datasets/pipelines/augmentations.py
@@ -164,10 +164,10 @@ class Imgaug:
                         shear=17.19 * cur_level * random.choice([-1, 1])),
                     dict(
                         type='TranslateX',
-                        px=100 * cur_level * random.choice([-1, 1])),
+                        px=int(100 * cur_level * random.choice([-1, 1]))),
                     dict(
                         type='TranslateY',
-                        px=100 * cur_level * random.choice([-1, 1])),
+                        px=int(100 * cur_level * random.choice([-1, 1]))),
                     dict(
                         type='Rotate',
                         rotate=30 * cur_level * random.choice([-1, 1])),
